ticket 410: emit-element no longer emits spurious newlines
This allows round-tripping (the data) of trivial XML examples,
provided that 408 has been fixed.

This also fixes #277 and is (now that I look at it) the same patch as
was originally submitted there by 'tjg'.

Whitespace in XML is problematic. We may be able to safely insert
whitespace (e.g. pretty printing) or not. This depends on wether the
whitespace at that point has any significance for consumers. This is
not possible when given an arbitary XML without knowledge of its
semantics and usage.