♯413 parse-args defaults in-enc and out-enc to UTF-8, as required by sh
Previously parse-args was defaulting in-enc and out-enc to the
platform default charset. This contradicted the intent of sh, which is
to default to UTF-8 on all platforms.

This appears not to have been noticed because the unit tests were
still testing for the previous behavior of defaulting to platform
encoding.

(As it turns out the old behavior of using Charset/defaultCharset
would have been wrong on Mac OS X since it claims "Mac Roman" here
despite the fact that Mac OS X uses UTF-8 throughout, including in
Terminal.app, shell and file system.)