CLJ-852: add special cases for "int"…"boolean" to tagToClass()
This causes the test to pass, however, there are still some open
questions:

Why does tagToClass first call maybeClass and then possibly
overwrite the result it gets back without even looking at it?
This strikes me as surprising. I would have expected a check for null,
and then only running the if/else chain of special cases if
maybeClass failed to return a non-null result.

The suppressed exception in maybeClass commented with "//aargh"
makes me wonder if maybeClass and tagToClass are factored correctly.